refactor: `src/pages/ReservationStatus` 경로 밑으로 `ReservationInformation` 컴포넌트를 이동하고 `useReservationInformation` 훅 분리
