+++ src/pages/ReservationStatus/components/reservationInformation.scss
