+++ src/pages/ReservationStatus/index.tsx
@@ -1,9 +1,9 @@
 import { useGetActivityListQuery, useLazyGetReservationDashboardQuery } from '@/api/reservationStatusApi';
 import CalendarReservationChip from '@/components/Calendar/components/CalendarReservationChip';
-import ReservationInformation from '@/components/Calendar/components/ReservationInformation';
 import useCalendar from '@/components/Calendar/hooks/useCalendar';
 import { DropDown, DropDownValue, DropdownItem } from '@/components/Dropdown';
 import { useModal } from '@/hooks/useModal/useModal';
+import ReservationInformation from '@/pages/ReservationStatus/components/ReservationInformation';
 import { isSameDate, statusChips, statusMap, translateMap } from '@/utils/calendarUtils';
 import { format, getYear } from 'date-fns';
 import { useEffect, useState } from 'react';
